# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ignite Backend (Prototype)
  version: 0.2.0
  description: >
    Authentication endpoints for registration with OTP, verification, resend,
    passwordless resume tokens, login, and token refresh.
servers:
  - url: http://localhost:5000/api
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /connections:
    get:
      tags: ["Connections"]
      summary: List user's connections
      description: |
        Get a paginated list of the authenticated user's connections with profile information.
        Returns connections sorted by most recent first.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of connections per page (max 100)
      responses:
        "200":
          description: Successfully retrieved connections
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConnectionProfile"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                unauthorized:
                  value:
                    code: UNAUTHORIZED
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                internal_error:
                  value:
                    code: INTERNAL
  /connections/requests/{id}/cancel:
    post:
      tags: ["Connections"]
      summary: Cancel a pending connection request
      description: Withdraw a pending request.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the connection request to cancel
      responses:
        "200":
          description: Connection request successfully canceled (or was already canceled)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                required:
                  - success
        "401":
          description: Authentication required or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                not_authenticated:
                  value:
                    code: NOT_AUTHENTICATED
        "404":
          description: Connection request not found or user is not the sender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                not_found:
                  value:
                    code: NOT_FOUND
        "409":
          description: Connection request cannot be canceled in its current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                invalid_state:
                  value:
                    code: INVALID_STATE
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                internal_error:
                  value:
                    code: INTERNAL
  /connections/{profileId}:
    delete:
      tags: ["Connections"]
      summary: Remove a connection
      description: Delete an existing connection between the authenticated user and the specified profile. This operation is idempotent - deleting a non-existing connection still returns success.
      parameters:
        - name: profileId
          in: path
          required: true
          description: The UUID of the profile to disconnect from
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connection successfully removed (or was already absent)
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
                    enum: [true]
              examples:
                success:
                  value:
                    success: true
        "401":
          description: Authentication required or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                not_authenticated:
                  value:
                    code: NOT_AUTHENTICATED
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorSimple"
              examples:
                internal_error:
                  value:
                    code: INTERNAL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorSimple:
      type: object
      required: [code]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - EMAIL_EXISTS
            - ZID_EXISTS
            - ZID_MISMATCH
            - PENDING_VERIFICATION_EXISTS
            - PENDING_NOT_FOUND
            - ALREADY_VERIFIED
            - RESUME_TOKEN_INVALID
            - OTP_INVALID
            - OTP_EXPIRED
            - OTP_LOCKED
            - OTP_COOLDOWN
            - OTP_RESEND_LIMIT
            - INVALID_CREDENTIALS
            - ACCOUNT_NOT_VERIFIED
            - NO_REFRESH_TOKEN
            - INVALID_REFRESH_TOKEN
            - NOT_AUTHENTICATED
            - USER_NOT_FOUND
            - NO_PENDING_EMAIL_CHANGE
            - TOO_MANY_REQUESTS
            - INTERNAL
            - HANDLE_TAKEN
            - FILE_TOO_LARGE              
            - UNSUPPORTED_MEDIA_TYPE      
            - INTERNAL_ERROR
            - NOT_FOUND
            - INVALID_STATE

        details:
          type: string
          description: Optional additional details about the error
...
    ConnectionProfile:
      type: object
      required: [profileId, fullName, handle, photoUrl, headline]
      properties:
        profileId:
          type: string
          format: uuid
          description: UUID of the connected user's profile
          example: "123e4567-e89b-12d3-a456-426614174000"
        fullName:
          type: string
          description: Full name of the connected user
          example: "Alex Johnson"
        handle:
          type: string
          nullable: true
          description: User's handle/username
          example: "alex"
        photoUrl:
          type: string
          nullable: true
          description: URL to the user's profile photo
          example: "https://example.com/photos/alex.jpg"
        headline:
          type: string
          nullable: true
          description: User's professional headline
          example: "Software Engineer at Tech Corp"
    PaginationInfo:
      type: object
      required: [total, page, pageSize, totalPages]
      properties:
        total:
          type: integer
          description: Total number of connections
          example: 50
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of items per page
          example: 20
        totalPages:
          type: integer
          description: Total number of pages
          example: 3
...